requirements.txt
flask
gunicorn
python-binance
binance-futures-connector
